Fix error if image is missing
Previously there would be an exception if no image is assigned and also
none of the other image options is selected.
Although form validation prevents that state from occurring, it can 
still happen due to on_delete=SET_NULL on the FilerImageField. So if a
referenced file in django-filer is deleted, form validation is bypassed.