Correct link to MDN's docs on the HTML img element
This file's comment section currently links to the MDN documentation page for the HTML image element. This element is not only non-standard and should not be used, but isn't actually used in the code itself. I changed the link to the MDN documentation page for the HTML img element. A simple, quick change to avoid any confusion for anyone trying to see how the code works.